seed Data를 통해 서버 구동시 DB 초기화
* 기존에 manuals라는 Collection이 있을시엔 삭제하고 새로 생성하도록 하였습니다.
* readdir을 별도 파일에서 처리하고 module.exports로 받아오고 싶었는데,
아무리 해도 결과를 받아올 수가 없어서 할 수 없이 한 파일 내에서 같이
처리하도록 하였습니다.

+++ backend/config/db.js
@@ -1,8 +1,35 @@
+const fs = require("fs");
 const mongoose = require("mongoose");
+const readline = require("readline");
 
 const { dbUri } = require("./index");
 const Manual = require("../models/manual");
-const manuals = require("../data/seed");
+
+let seed = null;
+
+fs.readdir(`seed`, async (err, files) => {
+  const readFile = file =>
+    new Promise((resolve, reject) => {
+      const rl = readline.createInterface({
+        input: fs.createReadStream(`seed/${file}`)
+      });
+      let lines = [];
+
+      rl
+        .on("line", line => {
+          lines.push(line.replace(/"/g, "").split(","));
+        })
+        .on("close", () => {
+          resolve(lines.join("\n"));
+        });
+    });
+
+  const promises = files.map(f => readFile(f));
+
+  seed = await Promise.all(promises)
+    .then(data => data.map(value => ({ value })))
+    .catch(console.error);
+});
 
 module.exports = () => {
   const connection = mongoose.connection;
@@ -10,26 +37,34 @@ module.exports = () => {
   mongoose.Promise = global.Promise;
 
   connection.on(`error`, console.error);
-  connection.once(`open`, () => {
-    console.log("MongoDB connect");
-  });
+  connection.once(`open`, () => console.log("MongoDB connect"));
 
   mongoose
     .connect(dbUri)
     .then(_ =>
-      connection.db.dropCollection("manuals", error => {
-        if (error) console.error(`dropCollection error ${error}`);
-      })
+      connection.db
+        .listCollections({ name: "manuals" })
+        .next((error, collectionInfo) => {
+          if (error) console.error(`listCollections next error ${error}`);
+
+          if (collectionInfo) {
+            connection.db.dropCollection("manuals", error => {
+              if (error) console.error(`dropCollection error ${error}`);
+            });
+          }
+        })
     )
     .then(_ => {
       Manual.counterReset("id", error => {
         if (error) console.error(`countRest Error ${error}`);
 
-        for (m of manuals)
-          Manual(m).save(error => {
+        console.log(seed);
+        for (s of seed) {
+          Manual(s).save(error => {
             if (error) console.error(`Document Create Error ${error}`);
           });
+        }
       });
     })
-    .catch(reason => console.error(reason));
+    .catch(reason => console.error(`Promise Fail ${reason}`));
 };
